Move unspent txo from transaction to seperate file.
