+++ lib/transaction.h
@@ -16,14 +16,4 @@ private:
     const std::vector<TxOut> txOuts_;
 };
 
-class UnspentTxOut {
-public:
-    UnspentTxOut(std::string txOutId, std::string address, size_t txOutIndex, size_t amount);
-private:
-    const std::string txOutId_;
-    const std::string address_;
-    const size_t txOutIndex_;
-    const size_t amount_;
-}
-
 #endif
